Background implemented and Texture Shift fix
Added the ability to change the Background Image in Level Script Editor, added RAM bank dumping, fixed texture shift parameter being incorrect for RCP_TEXGEN and RCP_TEXGEN_LINEAR in the model importer.

+++ SM64RAM/Utility.cs
@@ -29,7 +29,7 @@ namespace SM64RAM
         public static void SetBit(ref long value, int position, bool bit)
         {
             value = (value & ~(1 << position));
-            if (bit) value |= (1 << position);
+            if (bit) value |= (long)(1 << position);
         }
         public static bool GetBit(long value, int position)
         {
